Split CSS modules because prod build seems to screw up with the standard nav + hamburger nav styles in the same file...
